+++ Changelog.md
@@ -1,5 +1,8 @@
 # TORG Eternity Changelog
 
+## NEXT
+
+- Get Active Effect "Transfer on Attack" flag working (as part of making our own Data Model for Active Effects).
 
 ## 13.10.0